backend/requirements.txt
Django==4.1
djangorestframework==3.14
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-decouple==3.7
django-cors-headers==3.13.0